Move internal testcases to _internal_test.go file
