+++ command/stream_test.go
@@ -3,7 +3,6 @@ package command_test
 import (
 	"bytes"
 	"fmt"
-	"os"
 	"testing"
 
 	base "github.com/hirakiuc/daily-workflow/command"
@@ -68,45 +67,4 @@ func TestPrinterPrintf(t *testing.T) {
 	str := buf.String()
 	expected := fmt.Sprintf(format, val)
 	assert.Equal(t, str, expected)
-}
-
-func TestNewIoStream(t *testing.T) {
-	t.Parallel()
-
-	s := base.NewIoStream()
-	assert.NotNil(t, s)
-
-	assert.NotNil(t, s.Out)
-	assert.Equal(t, s.Out.Out, os.Stdout)
-
-	assert.NotNil(t, s.Err)
-	assert.Equal(t, s.Err.Out, os.Stderr)
-}
-
-func TestIoStreamSetOut(t *testing.T) {
-	t.Parallel()
-
-	s := base.NewIoStream()
-	assert.NotNil(t, s)
-
-	var buf bytes.Buffer
-
-	assert.NotEqual(t, s.Out.Out, &buf)
-
-	s.SetOut(&buf)
-	assert.Equal(t, s.Out.Out, &buf)
-}
-
-func TestIoStreamSetErr(t *testing.T) {
-	t.Parallel()
-
-	s := base.NewIoStream()
-	assert.NotNil(t, s)
-
-	var buf bytes.Buffer
-
-	assert.NotEqual(t, s.Err.Out, &buf)
-
-	s.SetErr(&buf)
-	assert.Equal(t, s.Err.Out, &buf)
 }
